Whenever running commands against the database make sure the DB object is created before attempting to use it's quote features
git-svn-id: http://svn.php.net/repository/pear/packages/Auth/trunk@207720 c90b9560-bf6c-de11-be94-00142212c4b1

+++ Container/MDB.php
@@ -343,6 +343,10 @@ class Auth_Container_MDB extends Auth_Container
      */
     function addUser($username, $password, $additional = "")
     {
+        $err = $this->_prepare();
+        if ($err !== true) {
+            return PEAR::raiseError($err->getMessage(), $err->getCode());
+        }
 
         if (isset($this->options['cryptType']) && $this->options['cryptType'] == 'none') {
             $cryptFunction = 'strval';
@@ -395,6 +399,11 @@ class Auth_Container_MDB extends Auth_Container
      */
     function removeUser($username)
     {
+        $err = $this->_prepare();
+        if ($err !== true) {
+            return PEAR::raiseError($err->getMessage(), $err->getCode());
+        }
+
         $query = sprintf("DELETE FROM %s WHERE %s = %s",
                          $this->options['table'],
                          $this->options['usernamecol'],
@@ -420,6 +429,11 @@ class Auth_Container_MDB extends Auth_Container
      */
     function changePassword($username, $password)
     {
+        $err = $this->_prepare();
+        if ($err !== true) {
+            return PEAR::raiseError($err->getMessage(), $err->getCode());
+        }
+
         if (isset($this->options['cryptType']) && $this->options['cryptType'] == 'none') {
             $cryptFunction = 'strval';
         } elseif (isset($this->options['cryptType']) && function_exists($this->options['cryptType'])) {
